refactor(app): #373 Phase 1-2 use-file-tabs.ts に editor / diff tab を切り出し
App.tsx 2686 -> 2341 行 (-345)。editor tab / diff tab / recentlyClosed
state と open*/save*/close*/cycleTab/togglePin/reopenLastClosed 等の
handler、dirtyEditorTabs / confirmDiscardEditorTabs の派生値を
src/renderer/src/lib/hooks/use-file-tabs.ts に集約。

Phase 1-1 の流儀に揃えて opts (projectRoot / refreshGit / gitStatus /
showToast) を optsRef.current = opts で毎 render 更新し、内部
useCallback の deps から外して TDZ 回避と再生成最小化を両立。
useT は hook 内で直接呼ぶ (settings-context.tsx の流儀)。

projectSwitchedRef.current の中で editor/diff/recentlyClosed/activeTabId
リセットは hook が expose する resetForProjectSwitch() に 1 行で委譲。
sessions / teams / terminal リセットは引き続き App.tsx の責務。

confirmDiscardRef / projectSwitchedRef の循環解消用 ref ブリッジは
useFileTabs と useProjectLoader の依存方向 (tabs が project を必要、
project が tabs.confirm を必要) を維持するため Phase 1-2 でも残す。
役割は縮小。

editor/diff には DnD が存在しない (DnD は terminal タブ専用) ため
本 PR では DnD を扱わず、Phase 1-3 (use-terminal-tabs.ts) に送る。

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy: 既存ベースライン 15 件のまま (新規警告ゼロ)

Refs #373
